refactor: PaymentRequest 클래스 위치 변경
PaymentRequest를 적절한 패키지로 이동하여 모듈 간 의존성을 정리했습니다.
