+++ src/main/java/roomescape/payment/domain/PaymentClient.java
@@ -1,6 +1,6 @@
 package roomescape.payment.domain;
 
-import roomescape.reservation.presentation.dto.PaymentRequest;
+import roomescape.payment.infrastructure.dto.PaymentRequest;
 import roomescape.payment.infrastructure.dto.PaymentResponse;
 
 public interface PaymentClient {
