+++ src/main/java/roomescape/payment/infrastructure/TossPaymentClient.java
@@ -2,7 +2,6 @@ package roomescape.payment.infrastructure;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import java.net.SocketTimeoutException;
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 
@@ -10,11 +9,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
-import roomescape.payment.exception.PaymentRequestException;
 import roomescape.payment.exception.RequestPaymentErrorHandler;
 import roomescape.payment.domain.PaymentClient;
-import roomescape.reservation.exception.ReservationException;
-import roomescape.reservation.presentation.dto.PaymentRequest;
+import roomescape.payment.infrastructure.dto.PaymentRequest;
 import roomescape.payment.infrastructure.dto.PaymentResponse;
 
 @Component
